Remove remote container binary cache
This container was likely previously being developed on a ubi8 base
image and populated the remote cache which those image layers.

We've updated the base image to ubi9 and it seems that the Dockerfile
instructions are still hitting the remote caches for layers built upon
the ubi8 base image.

This caused all sorts of issues as there were references in those cached
layers that were no longer the case with the ubi9 base image (different
gcc version, spack modulefile path was different, etc.).

Signed-off-by: Anderson Chauphan <achauph@sandia.gov>