Change permissions to avoid doubling image size
This approach creates files under /home/runner/ to all so that we don't need to
change permissions at the end and double the size of the Docker image!

And this also reduces the size of the image created by more than 2x. Apparently,
the Dockerfile before this was changing the permissions for large numbers of
files between layers and was more than doubling the size of the generated image.

Signed-off-by: Roscoe A. Bartlett <rabartl@sandia.gov>