Add googletest spack package
Trilinos PR#14719 removes Gtest as an internal Trilinos package and
instead requiring users to have a local install ready to be used,
similar how we handle other TPLs.

This PR aims to add Gtest to our container environments so that users
hoping to have an all-in-one solution with our containers won't have to
manually install gtest.

Signed-off-by: Anderson Chauphan <achauph@sandia.gov>